feat: переписал на Neo4j

+++ src/main/java/org/scuni/userservice/controller/UserController.java
@@ -18,7 +18,7 @@ public class UserController {
 
     @PostMapping("/user")
     public ResponseEntity<Object> createUser(@RequestBody UserCreateDto userCreateDto) {
-        Integer userId = userService.createUser(userCreateDto);
+        Long userId = userService.createUser(userCreateDto);
         return ResponseEntity.status(HttpStatus.CREATED).body(
                 Map.of("userId", userId,
                         "imageUrl", "/api/v1/images/user/" + userId,
@@ -28,27 +28,27 @@ public class UserController {
     }
 
     @GetMapping("/user/{id}")
-    public ResponseEntity<UserReadDto> getUserById(@PathVariable("id") Integer id) {
+    public ResponseEntity<UserReadDto> getUserById(@PathVariable("id") Long id) {
         UserReadDto user = userService.getUserById(id);
         return ResponseEntity.status(HttpStatus.OK)
                 .body(user);
     }
 
     @DeleteMapping("user/{id}")
-    public ResponseEntity<?> deleteUserById(@PathVariable("id") Integer id) {
+    public ResponseEntity<?> deleteUserById(@PathVariable("id") Long id) {
         userService.deleteUserById(id);
         return ResponseEntity.status(HttpStatus.NO_CONTENT)
                 .build();
     }
 
     @PutMapping("user/{id}/update_password")
-    public ResponseEntity<?> updatePassword(@PathVariable("id") Integer id) {
+    public ResponseEntity<?> updatePassword(@PathVariable("id") Long id) {
         userService.updatePassword(id);
         return ResponseEntity.status(HttpStatus.OK).build();
     }
 
     @PutMapping("user/{id}/send-verification-email")
-    public ResponseEntity<?> sendVerificationEmail(@PathVariable("id") Integer id) {
+    public ResponseEntity<?> sendVerificationEmail(@PathVariable("id") Long id) {
         userService.sendVerificationEmail(id);
         return ResponseEntity.status(HttpStatus.OK).build();
     }
